feat: implement real time call quality metrics
